+++ src/components/UI/Global/Styles.tsx
@@ -17,7 +17,7 @@ export default createGlobalStyle`
     background: ${props => props.theme.ui.colors.background.foundation};
     color: ${props => props.theme.ui.colors.text.primary};
     border: 1px solid ${props => props.theme.ui.colors.border.light};
-    border-radius: 10px;
+    border-radius: 5px;
     box-shadow: 0 2.5px 5px rgba(0, 0, 0, 0.2);
     overflow: hidden;
   }